Setup Project
I have completed the following tasks:

Created a new GitHub repository.

Repository name: my-project
Repository URL: https://github.com/DevKaranJ/Hello-Microverse
Created a new branch according to the GitHub flow rules.

Branch name: feature/setup-project
Added a .gitignore file to the repository.

Ignored the node_modules directory to exclude it from version control.
Set up a linter for HTML and CSS.

Ensured that any irrelevant files are ignored by Git.
Added a descriptive README file to the project.

Used the template provided and customized it for the project.
The README file provides information about the project, its purpose, and how to use it.
Added three files to the repository in separate commits:

Committed a .html file that includes a header with the text "Hello Microverse!" using the <h1> tag.

File name: index.html
Committed a .css file that includes one class with styles for the header.

File name: style.css
Committed a test.md file with the text "This file should be ignored by git".

File name: test.md
Configured Git to ignore this file to ensure it is not present in the GitHub repository.
Opened a pull request.

Pull request title: Add header and styles
Provided a description of the changes made in the pull request.